+++ app/itam/viewsets/software_version.py
@@ -2,8 +2,6 @@ from rest_framework.reverse import reverse
 
 from drf_spectacular.utils import extend_schema, extend_schema_view, OpenApiParameter, OpenApiResponse
 
-# THis import only exists so that the migrations can be created
-from itam.models.software_version_history import SoftwareVersionHistory    # pylint: disable=W0611:unused-import
 from itam.serializers.software_version import (    # pylint: disable=W0611:unused-import
     Software,
     SoftwareVersion,