+++ app/project_management/viewsets/project_milestone.py
@@ -2,8 +2,6 @@ from drf_spectacular.utils import extend_schema, extend_schema_view, OpenApiPara
 
 from api.viewsets.common.tenancy import ModelViewSet
 
-# This import only exists so that the migrations can be created
-from project_management.models.project_milestone_history import ProjectMilestoneHistory    # pylint: disable=W0611:unused-import
 from project_management.serializers.project_milestone import (    # pylint: disable=W0611:unused-import
     ProjectMilestone,
     ProjectMilestoneModelSerializer,