+++ Release-Notes.md
@@ -12,6 +12,12 @@
 
     - Ticket Task model added behind feature flag `2025-00006` and will remain until production ready.
 
+- Removed Models:
+
+    - [see v1.19.0] Removed Depreciated "model History" models
+
+    - [see v1.19.0] Removed Depreciated "model Notes" models
+
 
 ## Version v1.27.0